[#84536] arm64: Increase max size of tb jump offsets
Previously, any jump that required an offset larger than the 16-bit
allocated to the `tb_jmp_offset` field would be truncated and mess up
the translation block patching.

The issue was detected specifically for arm64 as it tends to emit more
instructions, thus increasing the length between two translation blocks
maxing out this 16-bit size limit. In reality, x86-64 supports 32-bit
offsets and arm64 supports 26-bit offsets.

+++ include/exec-all.h
@@ -139,8 +139,8 @@ struct TranslationBlock {
 
     /* the following data are used to directly call another TB from
        the code of this one. */
-    uint16_t tb_next_offset[2]; /* offset of original jump target */
-    uint16_t tb_jmp_offset[2];  /* offset of jump instruction */
+    uint32_t tb_next_offset[2]; /* offset of original jump target */
+    uint32_t tb_jmp_offset[2];  /* offset of jump instruction */
     /* list of TBs jumping to this one. This is a circular list using
        the two least significant bits of the pointers to tell what is
        the next pointer: 0 = jmp_next[0], 1 = jmp_next[1], 2 =
